shhh now it works :)
Former-commit-id: f4193f27c8b9d3a9452d98aaf7000bcf631dfe99 [formerly 1c2eb9793a6153696e7c559e285f300b587b05fa]
Former-commit-id: 0b6214c528595a9d11577ca3d8e60d4ac14f5307

+++ src/librapid/array/denseStorage.hpp
@@ -15,7 +15,7 @@ namespace librapid::memory {
 
 		DenseStorage() = default;
 
-		explicit DenseStorage(int64_t size) :
+		explicit DenseStorage(size_t size) :
 				m_size(roundTo(size, Vc::Vector<T>::size())), m_heap(memory::malloc<T, d>(m_size)),
 				m_refCount(new std::atomic<int64_t>(1)) {
 #if defined(LIBRAPID_HAS_CUDA)
